api: cleanup before new api.
