+++ Classes/Algorithms/Holders/JWTAlgorithmDataHolderChain.h
@@ -15,12 +15,12 @@
 
 #pragma mark - Initialization
 - (instancetype)initWithHolders:(NSArray *)holders;
-- (instancetype)initWithHolder:(id<JWTAlgorithmDataHolder>)holder;
+- (instancetype)initWithHolder:(id<JWTAlgorithmDataHolderProtocol>)holder;
 
 #pragma mark - Appending
 - (instancetype)chainByAppendingChain:(JWTAlgorithmDataHolderChain *)chain;
 - (instancetype)chainByAppendingHolders:(NSArray *)holders;
-- (instancetype)chainByAppendingHolder:(id<JWTAlgorithmDataHolder>)holder;
+- (instancetype)chainByAppendingHolder:(id<JWTAlgorithmDataHolderProtocol>)holder;
 
 @end
